requirements.txt
requests>=2.28.0
python-dotenv>=0.19.0
lxml>=4.9.0 